+++ api/data/networks.js
@@ -199,7 +199,7 @@ module.exports = [
   {
     id: 'kava-mainnet',
     title: 'Kava',
-    chain_id: 'kava-2',
+    chain_id: 'kava-3',
     api_url: 'http://lcd.kava.forbole.com:1317',
     rpc_url: 'ws://rpc.kava.forbole.com:26657/websocket',
     bech32_prefix: 'kava',
@@ -207,7 +207,7 @@ module.exports = [
     address_creator: 'cosmos',
     ledger_app: 'cosmos',
     network_type: 'cosmos',
-    source_class_name: 'source/kavaV0-source',
+    source_class_name: 'source/kavaV1-source',
     block_listener_class_name: 'block-listeners/cosmos-node-subscription',
     testnet: false,
     ...getNetworkCapabilities[`kava-mainnet`],
